+++ OcsStore/Models/CustomerTransactionView.cs
@@ -13,7 +13,7 @@ public partial class CustomerTransactionView
 
     public int MainId { get; set; }
 
-    public int Customer { get; set; }
+    public short Customer { get; set; }
 
     public decimal Amount { get; set; }
